Move new protocol event types
