+++ crates/espresso/node/src/lib.rs
@@ -838,6 +838,7 @@ pub mod testing {
         data::EpochNumber,
         event::LeafInfo,
         light_client::StateKeyPair,
+        new_protocol::CoordinatorEvent,
         signature_key::BLSKeyPair,
         traits::{
             EncodeBytes, block_contents::BlockHeader, metrics::NoMetrics, network::Topic,
@@ -855,7 +856,6 @@ pub mod testing {
     use super::*;
     use crate::{
         catchup::ParallelStateCatchup,
-        consensus_handle::CoordinatorEvent,
         persistence::no_storage::{self, NoStorage},
     };
 
@@ -1605,13 +1605,13 @@ mod test {
     use hotshot_example_types::node_types::TEST_VERSIONS;
     use hotshot_types::{
         event::LeafInfo,
+        new_protocol::CoordinatorEvent,
         traits::block_contents::{BlockHeader, BlockPayload},
     };
     use testing::{TestConfigBuilder, wait_for_decide_on_handle};
 
     use self::testing::run_test_builder;
     use super::*;
-    use crate::consensus_handle::CoordinatorEvent;
 
     #[test_log::test(tokio::test(flavor = "multi_thread"))]
     async fn test_skeleton_instantiation() {
