+++ crates/hotshot/new-protocol/src/message.rs
@@ -1,21 +1,17 @@
 use std::marker::PhantomData;
 
 use committable::{Commitment, Committable};
+pub use hotshot_types::new_protocol::Proposal;
 use hotshot_types::{
-    data::{
-        EpochNumber, Leaf2, QuorumProposal2, QuorumProposalWrapper, VidDisperseShare2,
-        ViewChangeEvidence2, ViewNumber,
-    },
-    drb::DrbResult,
+    data::{EpochNumber, VidDisperseShare2, ViewNumber},
     message::Proposal as SignedProposal,
     request_response::ProposalRequestPayload,
     simple_certificate::{
-        LightClientStateUpdateCertificateV2, OneHonestThreshold, QuorumCertificate2,
-        SimpleCertificate, SuccessThreshold, TimeoutCertificate2, UpgradeCertificate,
+        OneHonestThreshold, SimpleCertificate, SuccessThreshold, TimeoutCertificate2,
     },
     simple_vote::{
-        CheckpointData, HasEpoch, LightClientStateUpdateVote2, QuorumData2, QuorumVote2,
-        SimpleVote, TimeoutData2, TimeoutVote2, Vote2Data,
+        CheckpointData, LightClientStateUpdateVote2, QuorumData2, QuorumVote2, SimpleVote,
+        TimeoutData2, TimeoutVote2, Vote2Data,
     },
     traits::{node_implementation::NodeType, signature_key::SignatureKey},
     vote::HasViewNumber,
@@ -29,102 +25,6 @@ pub type Certificate1<T> = SimpleCertificate<T, QuorumData2<T>, SuccessThreshold
 pub type Certificate2<T> = SimpleCertificate<T, Vote2Data<T>, SuccessThreshold>;
 pub type TimeoutCertificate<T> = SimpleCertificate<T, TimeoutData2, SuccessThreshold>;
 pub type TimeoutOneHonest<T> = SimpleCertificate<T, TimeoutData2, OneHonestThreshold>;
-
-/// Proposal to append a block.
-#[derive(Debug, Serialize, Deserialize, Clone, Eq, PartialEq, Hash)]
-#[serde(bound(deserialize = ""))]
-pub struct Proposal<T: NodeType> {
-    /// The block header to append
-    pub block_header: T::BlockHeader,
-
-    /// view number for the proposal
-    pub view_number: ViewNumber,
-
-    /// The epoch number corresponding to the block number.
-    ///
-    /// Can be `None` for pre-epoch version.
-    pub epoch: EpochNumber,
-
-    /// certificate that the proposal is chaining from
-    pub justify_qc: QuorumCertificate2<T>,
-
-    /// certificate proving the last block of the epoch is decided
-    pub next_epoch_justify_qc: Option<Certificate2<T>>,
-
-    /// Possible upgrade certificate, which the leader may optionally attach.
-    pub upgrade_certificate: Option<UpgradeCertificate<T>>,
-
-    /// Possible timeout certificate.
-    ///
-    /// If the `justify_qc` is not for a proposal in the immediately preceding
-    /// view, then a timeout certificate must be attached.
-    pub view_change_evidence: Option<TimeoutCertificate<T>>,
-
-    /// The DRB result for the next epoch.
-    ///
-    /// This is required only for the last block of the epoch. Nodes will verify
-    /// that it's consistent with the result from their computations.
-    #[serde(with = "serde_bytes")]
-    pub next_drb_result: Option<DrbResult>,
-
-    /// The light client state update certificate for the next epoch.
-    /// This is required for the epoch root.
-    pub state_cert: Option<LightClientStateUpdateCertificateV2<T>>,
-}
-
-impl<T: NodeType> HasViewNumber for Proposal<T> {
-    fn view_number(&self) -> ViewNumber {
-        self.view_number
-    }
-}
-
-impl<T: NodeType> HasEpoch for Proposal<T> {
-    fn epoch(&self) -> Option<EpochNumber> {
-        Some(self.epoch)
-    }
-}
-
-impl<T: NodeType> From<QuorumProposalWrapper<T>> for Proposal<T> {
-    fn from(wrapper: QuorumProposalWrapper<T>) -> Self {
-        let qp = wrapper.proposal;
-        Self {
-            block_header: qp.block_header,
-            view_number: qp.view_number,
-            epoch: qp.epoch.unwrap_or(EpochNumber::new(0)),
-            justify_qc: qp.justify_qc,
-            next_epoch_justify_qc: None,
-            upgrade_certificate: qp.upgrade_certificate,
-            view_change_evidence: qp.view_change_evidence.and_then(|e| match e {
-                ViewChangeEvidence2::Timeout(tc) => Some(tc),
-                ViewChangeEvidence2::ViewSync(_) => None,
-            }),
-            next_drb_result: qp.next_drb_result,
-            state_cert: qp.state_cert,
-        }
-    }
-}
-
-impl<T: NodeType> From<Proposal<T>> for QuorumProposalWrapper<T> {
-    fn from(p: Proposal<T>) -> Self {
-        QuorumProposalWrapper::from(QuorumProposal2 {
-            block_header: p.block_header,
-            view_number: p.view_number,
-            epoch: Some(p.epoch),
-            justify_qc: p.justify_qc,
-            next_epoch_justify_qc: None,
-            upgrade_certificate: p.upgrade_certificate,
-            view_change_evidence: p.view_change_evidence.map(ViewChangeEvidence2::Timeout),
-            next_drb_result: p.next_drb_result,
-            state_cert: p.state_cert,
-        })
-    }
-}
-
-impl<T: NodeType> From<Proposal<T>> for Leaf2<T> {
-    fn from(p: Proposal<T>) -> Self {
-        Self::from_quorum_proposal(&QuorumProposalWrapper::from(p))
-    }
-}
 
 #[derive(Clone, Copy, Debug, Eq, PartialEq, PartialOrd, Ord, Hash, Deserialize)]
 pub enum Unchecked {}
