+++ crates/hotshot/types/src/new_protocol/mod.rs
@@ -1,0 +1,7 @@
+//! types for the new HotShot protocol.
+
+pub mod event;
+pub mod proposal;
+
+pub use event::{CoordinatorEvent, NewDecideEvent};
+pub use proposal::Proposal;
