+++ slow-tests/tests/restart_tests.rs
@@ -32,7 +32,6 @@ use espresso_node::{
         self, data_source::testing::TestableSequencerDataSource, options::Query,
         test_helpers::STAKE_TABLE_CAPACITY_FOR_TEST,
     },
-    consensus_handle::CoordinatorEvent,
     context::SequencerContext,
     genesis::{Genesis, L1Finalized, StakeTableConfig},
     keyset::KeySet,
@@ -65,6 +64,7 @@ use hotshot_types::{
     event::{Event, EventType},
     light_client::StateKeyPair,
     network::{Libp2pConfig, NetworkConfig},
+    new_protocol::CoordinatorEvent,
     signature_key::{BLSPrivKey, BLSPubKey},
     traits::signature_key::SignatureKey,
     x25519,
